API refactored convolution primitive into backend.
