+++ kymatio/scattering1d/backend/torch_backend.py
@@ -1,6 +1,7 @@
 # Authors: Edouard Oyallon, Joakim Anden, Mathieu Andreux
 
 import torch
+import torch.nn as nn
 import torch.nn.functional as F
 
 from collections import namedtuple
@@ -117,6 +118,25 @@ def ifft(x):
 def concatenate_1d(x):
     return concatenate(x, -2)
 
+class Conv1DFFT(nn.Module):
+    def __init__(self, backend):
+        super(Conv1DFFT, self).__init__()
+        self.name = 'torch'
+        self.backend = backend
+
+    def preprocess_signal(self, x):
+        x_hat = self.backend.rfft(x)
+        return x_hat
+
+    def convolution(self, x_hat, conv_filter, subsampling_factor, domain=''):
+        y_c = self.backend.cdgmm(x_hat, conv_filter)
+        y_hat = self.backend.subsample_fourier(y_c, 2 ** subsampling_factor)
+        if domain == 'real':
+            y_r = self.backend.irfft(y_hat)
+        else:
+            y_r = self.backend.ifft(y_hat)
+        return y_r
+
 
 backend = namedtuple('backend', ['name', 'modulus', 'subsample_fourier', 'unpad', 'fft', 'concatenate'])
 backend.name = 'torch'
@@ -129,3 +149,4 @@ backend.rfft = rfft
 backend.irfft = irfft
 backend.ifft = ifft
 backend.concatenate = concatenate_1d
+backend.conv = Conv1DFFT(backend)
