+++ kymatio/scattering1d/frontend/base_frontend.py
@@ -7,7 +7,6 @@ import numpy as np
 from ..filter_bank import scattering_filter_factory
 from ..utils import (compute_border_indices, compute_padding, compute_minimum_support_to_pad,
 compute_meta_scattering, precompute_size_scattering)
-from ..convolution import Conv1DFFT
 
 
 class ScatteringBase1D(ScatteringBase):
@@ -69,7 +68,6 @@ class ScatteringBase1D(ScatteringBase):
         self.ind_start, self.ind_end = compute_border_indices(
             self.J, self.pad_left, self.pad_left + self.T)
 
-        self.convolution = Conv1DFFT(self.backend)
 
     def create_filters(self):
         # Create the filters
