+++ kymatio/scattering1d/frontend/torch_frontend.py
@@ -6,9 +6,8 @@ import warnings
 
 from ...frontend.torch_frontend import ScatteringTorch
 from ..core.scattering1d import scattering1d
-from ..utils import precompute_size_scattering
+from ..utils import precompute_size_scattering, check_hermitian_symmetric
 from .base_frontend import ScatteringBase1D
-
 
 
 class ScatteringTorch1D(ScatteringTorch, ScatteringBase1D):
@@ -27,9 +26,13 @@ class ScatteringTorch1D(ScatteringTorch, ScatteringBase1D):
         will create the filters as numpy array, and then, it
         saves those arrays as module's buffers."""
         n = 0
+        self.phi_f_dict = dict()
+        self.psi1_f_dict = dict()
+        self.psi2_f_dict = dict()
         # prepare for pytorch
         for k in self.phi_f.keys():
             if type(k) != str:
+                self.phi_f_dict[n] = check_hermitian_symmetric(self.phi_f[k])
                 # view(-1, 1).repeat(1, 2) because real numbers!
                 self.phi_f[k] = torch.from_numpy(
                         self.phi_f[k]).float().view(-1, 1)
@@ -38,6 +41,7 @@ class ScatteringTorch1D(ScatteringTorch, ScatteringBase1D):
         for psi_f in self.psi1_f:
             for sub_k in psi_f.keys():
                 if type(sub_k) != str:
+                    self.psi1_f_dict[n] = check_hermitian_symmetric(psi_f[sub_k])
                     # view(-1, 1).repeat(1, 2) because real numbers!
                     psi_f[sub_k] = torch.from_numpy(
                         psi_f[sub_k]).float().view(-1, 1)
@@ -46,6 +50,8 @@ class ScatteringTorch1D(ScatteringTorch, ScatteringBase1D):
         for psi_f in self.psi2_f:
             for sub_k in psi_f.keys():
                 if type(sub_k) != str:
+                    print(check_hermitian_symmetric(psi_f[sub_k]))
+                    self.psi2_f_dict[n] = check_hermitian_symmetric(psi_f[sub_k])
                     # view(-1, 1).repeat(1, 2) because real numbers!
                     psi_f[sub_k] = torch.from_numpy(
                             psi_f[sub_k]).float().view(-1, 1)
@@ -119,7 +125,9 @@ class ScatteringTorch1D(ScatteringTorch, ScatteringBase1D):
                        oversampling=self.oversampling,
                        vectorize=self.vectorize,
                        size_scattering=size_scattering,
-                       out_type=self.out_type, conv_type=self.convolution)
+                       out_type=self.out_type,
+                       psi1_f_dict=self.psi1_f_dict, psi2_f_dict=self.psi2_f_dict, 
+                       phi_f_dict=self.phi_f_dict)
 
         if self.out_type == 'array' and self.vectorize:
             scattering_shape = S.shape[-2:]
